Added resolve-hostnames example program.
