+++ SDL_net.c
@@ -1465,63 +1465,3 @@ int SDLNet_WaitUntilInputAvailable(void **vsockets, int numsockets, int timeoutm
 
     return retval;
 }
-
-
-#if 0  // some test code.
-#include <stdio.h>
-
-int main(int argc, char **argv)
-{
-    if (SDLNet_Init() < 0) {
-        SDL_Log("SDLNet_Init() failed: %s", SDL_GetError());
-        return 1;
-    }
-
-#if 1
-    SDLNet_StreamSocket *stream;
-    if (argc > 1) {
-        SDLNet_Server *server = SDLNet_CreateServer(NULL, 7997);
-        SDLNet_WaitUntilConnected(server);
-        SDLNet_AcceptClient(server, &stream);
-    } else {
-        SDLNet_Address *addr = SDLNet_ResolveHostname("localhost");
-        SDLNet_WaitUntilResolved(addr, -1);
-        stream = SDLNet_CreateClient(addr, 7997);
-        SDLNet_WaitUntilConnected(stream, -1);
-    }
-
-    printf("\n\nConnected!\n\n");
-
-    fcntl(fileno(stdin), F_SETFL, fcntl(fileno(stdin), F_GETFL, 0) | O_NONBLOCK);
-
-    while (1) {
-        char buf[128];
-        buf[0] = '\0';
-        fgets(buf, sizeof (buf), stdin);
-        if (buf[0]) {
-            SDLNet_WriteToStreamSocket(stream, buf, SDL_strlen(buf));
-        }
-        const int br = SDLNet_ReadStreamSocket(stream, buf, sizeof (buf));
-        if (br > 0) {
-            fwrite(buf, 1, br, stdout);
-        }
-    }
-#endif
-
-#if 0
-    //SDLNet_SimulateAddressResolutionLoss(3000, 30);
-
-    SDLNet_Address **addrs = (SDLNet_Address **) SDL_calloc(argc, sizeof (SDLNet_Address *));
-    for (int i = 1; i < argc; i++) {
-        addrs[i] = SDLNet_ResolveHostname(argv[i]);
-    }
-
-    for (int i = 1; i < argc; i++) {
-        SDLNet_WaitUntilResolved(addrs[i], -1);
-    }
-#endif
-
-    SDLNet_Quit();
-    return 0;
-}
-#endif
